+++ contracts/factories/CompanyConfiguratorFactory.sol
@@ -50,23 +50,23 @@ contract CompanyConfiguratorFactory {
     uint64 minimumVotingTime = uint64(7 days);
 
     company.setVotingBylaw("setEntityStatus(address,uint8)", 1, 2, true, minimumVotingTime, favor);
-    company.setStatusBylaw("beginPoll(address,uint64,bool,bool)", uint(AbstractCompany.SpecialEntityStatus.Shareholder), true);
+    company.setStatusBylaw("beginPoll(address,uint64,bool,bool)", uint8(AbstractCompany.SpecialEntityStatus.Shareholder), true);
     company.setVotingBylaw("addStock(address,uint256)", 1, 2, true, minimumVotingTime, favor);
     company.setVotingBylaw("issueStock(uint8,uint256)", 1, 2, true, minimumVotingTime, favor);
-    company.setStatusBylaw("grantStock(uint8,uint256,address)", uint(AbstractCompany.EntityStatus.Executive), false);
+    company.setStatusBylaw("grantStock(uint8,uint256,address)", uint8(AbstractCompany.EntityStatus.Executive), false);
     company.setAddressBylaw("grantVestedStock(uint8,uint256,address,uint64,uint64,uint64)", msg.sender, false);
 
     company.setVotingBylaw("beginSale(address)", 1, 2, true, minimumVotingTime, favor);
-    company.setStatusBylaw("transferSaleFunds(uint256)", uint(AbstractCompany.EntityStatus.Executive), false);
+    company.setStatusBylaw("transferSaleFunds(uint256)", uint8(AbstractCompany.EntityStatus.Executive), false);
 
     company.setVotingBylaw("setAccountingSettings(uint256,uint64,uint256)", 1, 2, true, minimumVotingTime, favor);
-    company.setStatusBylaw("createRecurringReward(address,uint256,uint64,string)", uint(AbstractCompany.EntityStatus.Executive), false);
+    company.setStatusBylaw("createRecurringReward(address,uint256,uint64,string)", uint8(AbstractCompany.EntityStatus.Executive), false);
 
-    company.setStatusBylaw("removeRecurringReward(uint)", uint(AbstractCompany.EntityStatus.Executive), false);
-    company.setStatusBylaw("issueReward(address,uint256,string)", uint(AbstractCompany.EntityStatus.Executive), false);
+    company.setStatusBylaw("removeRecurringReward(uint)", uint8(AbstractCompany.EntityStatus.Executive), false);
+    company.setStatusBylaw("issueReward(address,uint256,string)", uint8(AbstractCompany.EntityStatus.Executive), false);
 
-    company.setStatusBylaw("assignStock(uint8,address,uint256)", uint(AbstractCompany.SpecialEntityStatus.StockSale), true);
-    company.setStatusBylaw("removeStock(uint8,address,uint256)", uint(AbstractCompany.SpecialEntityStatus.StockSale), true);
+    company.setStatusBylaw("assignStock(uint8,address,uint256)", uint8(AbstractCompany.SpecialEntityStatus.StockSale), true);
+    company.setStatusBylaw("removeStock(uint8,address,uint256)", uint8(AbstractCompany.SpecialEntityStatus.StockSale), true);
 
     // Protect bylaws under a 2/3 voting
     company.setVotingBylaw("setAddressBylaw(string,address,bool)", 2, 3, false, minimumVotingTime, favor);
